requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
loguru==0.7.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
oauthlib==3.2.2
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pip==24.2
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytimeparse==1.1.8
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==1.3.1
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.0.0
tornado==6.4.2
traitlets==5.14.3
tweepy==4.14.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.44.0
